requirements.txt
numpy
faiss-cpu
torch
sentence-transformers
accelerate
transformers
PyPDF2
hf_xet



#multimodal RAG
fitz
Pytesseract
chromadb
clip
easyocr
docx
PyMuPDF
bitsandbytes-windows


#app
streamlit
pydicom

#report 
reportlab
